Add support for actions and input fields
This adds some support for the following actions:
* Selection
* Text
* Button/ Action
To get the values from the actions and input fields some handlers were added.
These will be called by windows and execute user supplied closures.
Note that the closures are `Send` to avoid a future breaking change
because of an update of the `windows`-crate.